shared array methods
