+++ src/array/sync.rs
@@ -1,18 +1,15 @@
 //! The `Array` Python class: metadata accessors and numpy-style reads.
 
 use crate::array::selection::PySelection;
+use crate::array::shared::array_metadata_accessors;
 use crate::array::util::PyChunkIndices;
-use crate::chunks::PyChunkGrid;
-use crate::codec::{PyCodecChain, PyCodecOptions};
+use crate::codec::PyCodecOptions;
 use crate::decoded_array::DecodedArray;
-use crate::dtype::PyDataType;
 use crate::error::ZarristaResult;
 use crate::node::PyNodePath;
 use crate::storage::PySyncStorage;
 use pyo3::prelude::*;
 use pyo3_bytes::PyBytes;
-use pythonize::pythonize;
-use pythonize::Result as PythonizeResult;
 use zarrs::array::Array;
 use zarrs::storage::ReadableWritableListableStorageTraits;
 
@@ -27,6 +24,9 @@ impl PyArray {
         Self { inner }
     }
 }
+
+// Metadata accessors shared with `PyAsyncArray`; see `array/shared.rs`.
+array_metadata_accessors!(PyArray);
 
 #[pymethods]
 impl PyArray {
@@ -52,51 +52,6 @@ impl PyArray {
     fn open(store: PySyncStorage, path: PyNodePath) -> ZarristaResult<Self> {
         let inner = Array::open(store.into(), path.as_str())?;
         Ok(Self::new(inner))
-    }
-
-    /// The array's user attributes as a dict.
-    #[getter]
-    fn attrs<'py>(&self, py: Python<'py>) -> PythonizeResult<Bound<'py, PyAny>> {
-        pythonize(py, self.inner.attributes())
-    }
-
-    #[getter]
-    fn chunk_grid(&self) -> PyChunkGrid {
-        self.inner.chunk_grid().clone().into()
-    }
-
-    #[getter]
-    fn codecs(&self) -> PyCodecChain {
-        self.inner.codecs().into()
-    }
-
-    /// The dimension names, if any were specified.
-    #[getter]
-    fn dimension_names(&self) -> &Option<Vec<Option<String>>> {
-        self.inner.dimension_names()
-    }
-
-    /// The Zarr data-type
-    #[getter]
-    fn dtype(&self) -> PyDataType {
-        self.inner.data_type().clone().into()
-    }
-
-    #[getter]
-    fn metadata<'py>(&self, py: Python<'py>) -> Bound<'py, PyAny> {
-        pythonize(py, self.inner.metadata()).unwrap()
-    }
-
-    /// The number of dimensions.
-    #[getter]
-    fn ndim(&self) -> usize {
-        self.inner.dimensionality()
-    }
-
-    /// The array's path in the store.
-    #[getter]
-    fn path(&self) -> &str {
-        self.inner.path().as_str()
     }
 
     /// Read a region of the array, using numpy-style basic indexing.
@@ -128,11 +83,5 @@ impl PyArray {
     ) -> ZarristaResult<Option<PyBytes>> {
         let encoded = self.inner.retrieve_encoded_chunk(chunk_indices.as_ref())?;
         Ok(encoded.map(|buf| PyBytes::new(buf.into())))
-    }
-
-    /// The array shape.
-    #[getter]
-    fn shape(&self) -> &[u64] {
-        self.inner.shape()
     }
 }
